test(gateway): split observer digest-budget cases under max-lines
